refactor: extract formatErrorMessage to eliminate test duplication
Extract the error formatting when-expression from General.kt into a
standalone formatErrorMessage() function in ErrorFormatting.kt. This
allows the test-runner to call the production function directly via
symlink instead of maintaining a duplicated extractErrorMessage() helper.

The test already covered Array type; all 8 tests pass against the
production function.
