+++ android/app/src/main/java/com/mattermost/helpers/push_notification/General.kt
@@ -1,7 +1,6 @@
 package com.mattermost.helpers.push_notification
 
 import com.facebook.react.bridge.ReadableMap
-import com.facebook.react.bridge.ReadableType
 import com.mattermost.helpers.Network
 import com.mattermost.helpers.PushNotificationDataRunnable
 import com.mattermost.helpers.ResolvePromise
@@ -19,18 +18,7 @@ internal suspend fun PushNotificationDataRunnable.Companion.fetch(serverUrl: Str
                     // Server may return "data" as a Map (normal error response), String
                     // (e.g. HTML error pages, proxy errors), or other types. Handle each
                     // ReadableType explicitly to avoid UnexpectedNativeTypeException.
-                    val dataType = if (response.hasKey("data")) response.getType("data") else ReadableType.Null
-                    val errorMessage = when (dataType) {
-                        ReadableType.Map -> {
-                            val error = response.getMap("data")
-                            "Unexpected code ${error?.getInt("status_code")} ${error?.getString("message")}"
-                        }
-                        ReadableType.String -> "Unexpected response: ${response.getString("data")}"
-                        ReadableType.Number -> "Unexpected response: ${response.getDouble("data")}"
-                        ReadableType.Boolean -> "Unexpected response: ${response.getBoolean("data")}"
-                        ReadableType.Array -> "Unexpected response: ${response.getArray("data")}"
-                        ReadableType.Null -> "Unexpected response: null"
-                    }
+                    val errorMessage = formatErrorMessage(response)
                     cont.resumeWith(Result.failure(IOException(errorMessage)))
                 } else {
                     cont.resumeWith(Result.success(response))
